Fix imports and move Handler to events module
