+++ nicegui/elements/carousel.py
@@ -2,9 +2,10 @@ from __future__ import annotations
 
 from typing import Any, Optional, Union, cast
 
+from nicegui.events import Handler
+
 from ..context import context
 from ..events import ValueChangeEventArguments
-from .handler import Handler
 from .mixins.disableable_element import DisableableElement
 from .mixins.value_element import ValueElement
 
@@ -13,7 +14,7 @@ class Carousel(ValueElement):
 
     def __init__(self, *,
                  value: Union[str, CarouselSlide, None] = None,
-                 on_value_change: Optional[Handler[ValueChangeEventArguments]],
+                 on_value_change: Optional[Handler[ValueChangeEventArguments]] = None,
                  animated: bool = False,
                  arrows: bool = False,
                  navigation: bool = False,
