--- nicegui/elements/handler.py
@@ -1,6 +1,0 @@
-from typing import Any, Callable, TypeVar, Union
-
-from typing_extensions import TypeAlias
-
-EventT = TypeVar('EventT')
-Handler: TypeAlias = Union[Callable[[EventT], Any], Callable[[], Any]]
