+++ nicegui/elements/keyboard.py
@@ -2,6 +2,8 @@ from typing import List, Literal, Optional
 
 from typing_extensions import Self
 
+from nicegui.events import Handler
+
 from ..binding import BindableProperty
 from ..element import Element
 from ..events import (
@@ -12,7 +14,6 @@ from ..events import (
     KeyEventArguments,
     handle_event,
 )
-from .handler import Handler
 
 
 class Keyboard(Element, component='keyboard.js'):
