+++ nicegui/elements/menu.py
@@ -1,9 +1,10 @@
 from typing import Optional, Union
 
+from nicegui.events import Handler
+
 from ..element import Element
 from ..events import ClickEventArguments
 from .context_menu import ContextMenu
-from .handler import Handler
 from .item import Item
 from .mixins.value_element import ValueElement
