+++ nicegui/elements/notification.py
@@ -3,10 +3,11 @@ from typing import Any, Dict, Literal, Optional, Union
 
 from typing_extensions import Self
 
+from nicegui.events import Handler
+
 from ..context import context
 from ..element import Element
 from ..events import UiEventArguments, handle_event
-from .handler import Handler
 
 NotificationPosition = Literal[
     'top-left',
