+++ nicegui/elements/scene.py
@@ -4,6 +4,8 @@ from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Union
 
 from typing_extensions import Self
 
+from nicegui.events import Handler
+
 from .. import binding
 from ..dataclasses import KWONLY_SLOTS
 from ..element import Element
@@ -14,7 +16,6 @@ from ..events import (
     SceneDragEventArguments,
     handle_event,
 )
-from .handler import Handler
 from .scene_object3d import Object3D
