+++ nicegui/elements/stepper.py
@@ -2,10 +2,11 @@ from __future__ import annotations
 
 from typing import Any, Optional, Union, cast
 
+from nicegui.events import Handler
+
 from ..context import context
 from ..element import Element
 from ..events import ValueChangeEventArguments
-from .handler import Handler
 from .mixins.disableable_element import DisableableElement
 from .mixins.icon_element import IconElement
 from .mixins.value_element import ValueElement
@@ -15,7 +16,7 @@ class Stepper(ValueElement):
 
     def __init__(self, *,
                  value: Union[str, Step, None] = None,
-                 on_value_change: Optional[Handler[ValueChangeEventArguments]],
+                 on_value_change: Optional[Handler[ValueChangeEventArguments]] = None,
                  keep_alive: bool = True,
                  ) -> None:
         """Stepper
