+++ nicegui/elements/tree.py
@@ -2,8 +2,9 @@ from typing import Any, Dict, Iterator, List, Literal, Optional, Set
 
 from typing_extensions import Self
 
+from nicegui.events import Handler
+
 from ..events import GenericEventArguments, ValueChangeEventArguments, handle_event
-from .handler import Handler
 from .mixins.filter_element import FilterElement
